# swagger.yaml
openapi: 3.0.0
info:
  title: Mutual Fund API
  version: 1.0.0
servers:
  - url: http://localhost:3001
paths:
  /subscriptions:
    get:
      summary: Get all subscriptions
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Subscription'
    post:
      summary: Create a new subscription
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Subscription'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
  /subscriptions/{id}:
    get:
      summary: Get a subscription by ID
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Subscription ID
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        '404':
          description: Subscription not found
    put:
      summary: Update a subscription by ID
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Subscription ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Subscription'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Subscription'
        '404':
          description: Subscription not found
    delete:
      summary: Delete a subscription by ID
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Subscription ID
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: Subscription not found
components:
  schemas:
    Subscription:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        amount:
          type: number
          format: double
        date:
          type: string
          format: date